+++ packages/typespec-rust/CHANGELOG.md
@@ -14,6 +14,10 @@
 
 * Added support for TypeSpec `duration` types. Numeric durations are emitted as their respective types. For ISO8601 they're emitted as `String` types.
 
+### Other Changes
+
+* Removed dependency on crate `async-std`.
+
 ## 0.5.1 (2024-12-19)
 
 ### Bugs Fixed